responses

+++ src/main/java/org/henick/lottoapi/model/Prize.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 @Getter
 public class Prize {
     int degree;
-    long winnersCount;
+    int winnersCount;
     double value;
 
-    public Prize(int degree, long winnersCount, double value) {
+    public Prize(int degree, int winnersCount, double value) {
         this.degree = degree;
         this.winnersCount = winnersCount;
         this.value = value;
